implemented a method to run the logic of the program. It is not finished yet

+++ C++ OOP/headers/Customer.h
@@ -115,7 +115,7 @@ public:
     // methods
     void printInfoCustomer();
     void purchaseCar(Car& car);
-    void addMoney(int amount);
+    void addMoney(float amount);
 };
 
 int Customer::carPurchased = 0;
